+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.1.2
+*****
+
+* Updated the ``Database.__str__`` and ``Database.__repr__`` methods.
+* Added the ``Database.__eq__`` method.
+
 
 0.1.1
 *****